backend-api/requirements.txt
fastapi
uvicorn
numpy
ultralytics
opencv-python
Pillow
